+++ index.html
@@ -51,6 +51,6 @@
       </div>
     </div>
   </div>
-  <script src="./src/index.js"></script>
+  <script type = "module" src="./src/index.js"></script>
 </body>
 </html>
