+++ src/blocks/place-card/__delete-icon/place-card__delete-icon.css
@@ -1,5 +1,5 @@
 .place-card__delete-icon {
-  background: url(../../../images/trash-icon.svg) center no-repeat;
+  background: url(./../../../images/trash-icon.svg) center no-repeat;
   background-color: transparent;
   border: none;
   width: 18px;
